+++ lua/mini/base16.lua
@@ -31,6 +31,7 @@
 ---     - 'glepnir/dashboard-nvim'
 ---     - 'glepnir/lspsaga.nvim'
 ---     - 'HiPhish/nvim-ts-rainbow2'
+---     - 'HiPhish/rainbow-delimiters.nvim'
 ---     - 'hrsh7th/nvim-cmp'
 ---     - 'justinmk/vim-sneak'
 ---     - 'kevinhwang91/nvim-ufo'
@@ -863,6 +864,7 @@ H.apply_palette = function(palette, use_cterm)
     hi('OutlineIndentOdd',       {fg=p.base05, bg=nil, attr=nil, sp=nil})
   end
 
+  -- TODO: Remove after 'mini.nvim' 0.10.0 release
   if H.has_integration('HiPhish/nvim-ts-rainbow2') then
     hi('TSRainbowBlue',   {fg=p.base0D, bg=nil, attr=nil, sp=nil})
     hi('TSRainbowCyan',   {fg=p.base0C, bg=nil, attr=nil, sp=nil})
@@ -871,6 +873,16 @@ H.apply_palette = function(palette, use_cterm)
     hi('TSRainbowRed',    {fg=p.base08, bg=nil, attr=nil, sp=nil})
     hi('TSRainbowViolet', {fg=p.base0E, bg=nil, attr=nil, sp=nil})
     hi('TSRainbowYellow', {fg=p.base0A, bg=nil, attr=nil, sp=nil})
+  end
+
+  if H.has_integration('HiPhish/rainbow-delimiters.nvim') then
+    hi('RainbowDelimiterBlue',   {fg=p.base0D, bg=nil, attr=nil, sp=nil})
+    hi('RainbowDelimiterCyan',   {fg=p.base0C, bg=nil, attr=nil, sp=nil})
+    hi('RainbowDelimiterGreen',  {fg=p.base0B, bg=nil, attr=nil, sp=nil})
+    hi('RainbowDelimiterOrange', {fg=p.base09, bg=nil, attr=nil, sp=nil})
+    hi('RainbowDelimiterRed',    {fg=p.base08, bg=nil, attr=nil, sp=nil})
+    hi('RainbowDelimiterViolet', {fg=p.base0E, bg=nil, attr=nil, sp=nil})
+    hi('RainbowDelimiterYellow', {fg=p.base0A, bg=nil, attr=nil, sp=nil})
   end
 
   if H.has_integration('hrsh7th/nvim-cmp') then